{"types":"C","runs":10}

+++ concsolepr.java
@@ -3,9 +3,9 @@ import acm.program.ConsoleProgram;
 public class concsolepr extends ConsoleProgram {
 	public void run(){
 	
-	 	int num1 = readInt("Enter num1 =");
-		int num2 = readInt("Enter num2 =");
-		int average =(num1+num2)/2;
+	 	double num1 = readInt("Enter num1 =");
+		double num2 = readInt("Enter num2 =");
+		double average =(num1+num2)/2;
 		print(average);
